removed more lightbox, changed way of including javascript files(for older elgg versions), fixed a bug in synonyms
